rename gateio to gate_io, add 1m warning for gate_io_perpetual
