+++ hummingbot/data_feed/candles_feed/candles_factory.py
@@ -1,7 +1,7 @@
 from hummingbot.data_feed.candles_feed.binance_perpetual_candles import BinancePerpetualCandles
 from hummingbot.data_feed.candles_feed.binance_spot_candles import BinanceSpotCandles
-from hummingbot.data_feed.candles_feed.gateio_perpetual_candles import GateioPerpetualCandles
-from hummingbot.data_feed.candles_feed.gateio_spot_candles import GateioSpotCandles
+from hummingbot.data_feed.candles_feed.gate_io_perpetual_candles import GateioPerpetualCandles
+from hummingbot.data_feed.candles_feed.gate_io_spot_candles import GateioSpotCandles
 from hummingbot.data_feed.candles_feed.kucoin_spot_candles.kucoin_spot_candles import KucoinSpotCandles
 
 
@@ -18,9 +18,9 @@ class CandlesFactory:
             return BinancePerpetualCandles(trading_pair, interval, max_records)
         elif connector == "binance":
             return BinanceSpotCandles(trading_pair, interval, max_records)
-        elif connector == "gateio":
+        elif connector == "gate_io":
             return GateioSpotCandles(trading_pair, interval, max_records)
-        elif connector == "gateio_perpetual":
+        elif connector == "gate_io_perpetual":
             return GateioPerpetualCandles(trading_pair, interval, max_records)
         elif connector == "kucoin":
             return KucoinSpotCandles(trading_pair, interval, max_records)
