+++ hummingbot/data_feed/candles_feed/gate_io_perpetual_candles/__init__.py
@@ -1,0 +1,3 @@
+from hummingbot.data_feed.candles_feed.gate_io_perpetual_candles.gate_io_perpetual_candles import GateioPerpetualCandles
+
+__all__ = ["GateioPerpetualCandles"]
