+++ hummingbot/data_feed/candles_feed/gate_io_perpetual_candles/constants.py
