+++ hummingbot/data_feed/candles_feed/gate_io_perpetual_candles/gate_io_perpetual_candles.py
@@ -9,11 +9,11 @@ from hummingbot.core.network_iterator import NetworkStatus, safe_ensure_future
 from hummingbot.core.web_assistant.connections.data_types import WSJSONRequest
 from hummingbot.core.web_assistant.ws_assistant import WSAssistant
 from hummingbot.data_feed.candles_feed.candles_base import CandlesBase
-from hummingbot.data_feed.candles_feed.gateio_spot_candles import constants as CONSTANTS
+from hummingbot.data_feed.candles_feed.gate_io_perpetual_candles import constants as CONSTANTS
 from hummingbot.logger import HummingbotLogger
 
 
-class GateioSpotCandles(CandlesBase):
+class GateioPerpetualCandles(CandlesBase):
     _logger: Optional[HummingbotLogger] = None
 
     @classmethod
@@ -24,10 +24,11 @@ class GateioSpotCandles(CandlesBase):
 
     def __init__(self, trading_pair: str, interval: str = "1m", max_records: int = 150):
         super().__init__(trading_pair, interval, max_records)
+        self.quanto_multiplier = None
 
     @property
     def name(self):
-        return f"gateio_spot_{self._trading_pair}"
+        return f"gate_io_perpetual_{self._trading_pair}"
 
     @property
     def rest_url(self):
@@ -53,6 +54,14 @@ class GateioSpotCandles(CandlesBase):
     def intervals(self):
         return CONSTANTS.INTERVALS
 
+    async def start_network(self):
+        """
+        This method starts the network and starts a task for listen_for_subscriptions.
+        """
+        await self.stop_network()
+        await self.get_exchange_trading_pair_quanto_multiplier()
+        self._listen_candles_task = safe_ensure_future(self.listen_for_subscriptions())
+
     async def check_network(self) -> NetworkStatus:
         rest_assistant = await self._api_factory.get_rest_assistant()
         await rest_assistant.execute_request(url=self.health_check_url,
@@ -62,28 +71,41 @@ class GateioSpotCandles(CandlesBase):
     def get_exchange_trading_pair(self, trading_pair):
         return trading_pair.replace("-", "_")
 
+    async def get_exchange_trading_pair_quanto_multiplier(self):
+        rest_assistant = await self._api_factory.get_rest_assistant()
+        data = await rest_assistant.execute_request(
+            url=self.rest_url + CONSTANTS.CONTRACT_INFO_URL.format(contract=self._ex_trading_pair),
+            throttler_limit_id=CONSTANTS.CONTRACT_INFO_URL
+        )
+        quanto_multiplier = float(data.get("quanto_multiplier"))
+        self.quanto_multiplier = quanto_multiplier
+        return quanto_multiplier
+
     async def fetch_candles(self,
                             start_time: Optional[int] = None,
                             end_time: Optional[int] = None,
                             limit: Optional[int] = 500):
         rest_assistant = await self._api_factory.get_rest_assistant()
-        params = {"currency_pair": self._ex_trading_pair, "interval": self.interval, "limit": limit}
+        params = {"contract": self._ex_trading_pair, "interval": self.interval, "limit": limit}
+        if start_time or end_time:
+            del params["limit"]
         if start_time:
-            params["from"] = start_time
+            params["from"] = str(start_time)
         if end_time:
-            params["to"] = end_time
+            params["to"] = str(end_time)
+
         candles = await rest_assistant.execute_request(url=self.candles_url,
                                                        throttler_limit_id=CONSTANTS.CANDLES_ENDPOINT,
                                                        params=params)
         new_hb_candles = []
         for i in candles:
-            timestamp_ms = i[0] + "000"
-            open = i[5]
-            high = i[3]
-            low = i[4]
-            close = i[2]
-            volume = i[6]
-            quote_asset_volume = i[1]
+            timestamp_ms = i.get("t") * 1e3
+            open = i.get("o")
+            high = i.get("h")
+            low = i.get("l")
+            close = i.get("c")
+            volume = i.get("v") * self.quanto_multiplier
+            quote_asset_volume = i.get("sum")
             # no data field
             n_trades = 0
             taker_buy_base_volume = 0
@@ -135,6 +157,8 @@ class GateioSpotCandles(CandlesBase):
             subscribe_candles_request: WSJSONRequest = WSJSONRequest(payload=payload)
 
             await ws.send(subscribe_candles_request)
+            if self.interval == '1m':
+                self.logger().warning("The 1m K-line on gateioperpetual is currently not very accurate due to discrepancies between the official ws and rs data...")
             self.logger().info("Subscribed to public klines...")
         except asyncio.CancelledError:
             raise
@@ -148,30 +172,32 @@ class GateioSpotCandles(CandlesBase):
     async def _process_websocket_messages(self, websocket_assistant: WSAssistant):
         async for ws_response in websocket_assistant.iter_messages():
             data: Dict[str, Any] = ws_response.data
-            if data.get("event") == "update" and data.get("channel") == "spot.candlesticks":
-                timestamp_ms = int(data["result"]["t"] + "000")
-                open = data["result"]["o"]
-                high = data["result"]["h"]
-                low = data["result"]["l"]
-                close = data["result"]["c"]
-                volume = data["result"]["v"]
-                quote_asset_volume = data["result"]["a"]
-                # no data field
-                n_trades = 0
-                taker_buy_base_volume = 0
-                taker_buy_quote_volume = 0
-                if len(self._candles) == 0:
-                    self._candles.append(np.array([timestamp_ms, open, high, low, close, volume,
-                                                   quote_asset_volume, n_trades, taker_buy_base_volume,
-                                                   taker_buy_quote_volume]))
-                    safe_ensure_future(self.fill_historical_candles())
-                elif timestamp_ms > int(self._candles[-1][0]):
-                    # TODO: validate also that the diff of timestamp == interval (issue with 30d interval).
-                    self._candles.append(np.array([timestamp_ms, open, high, low, close, volume,
-                                                   quote_asset_volume, n_trades, taker_buy_base_volume,
-                                                   taker_buy_quote_volume]))
-                elif timestamp_ms == int(self._candles[-1][0]):
-                    self._candles.pop()
-                    self._candles.append(np.array([timestamp_ms, open, high, low, close, volume,
-                                                   quote_asset_volume, n_trades, taker_buy_base_volume,
-                                                   taker_buy_quote_volume]))
+
+            if data.get("event") == "update" and data.get("channel") == "futures.candlesticks":
+                for i in data["result"]:
+                    timestamp_ms = int(i["t"] * 1e3)
+                    open = i["o"]
+                    high = i["h"]
+                    low = i["l"]
+                    close = i["c"]
+                    volume = i["v"] * self.quanto_multiplier
+                    # no data field
+                    quote_asset_volume = 0
+                    n_trades = 0
+                    taker_buy_base_volume = 0
+                    taker_buy_quote_volume = 0
+                    if len(self._candles) == 0:
+                        self._candles.append(np.array([timestamp_ms, open, high, low, close, volume,
+                                                       quote_asset_volume, n_trades, taker_buy_base_volume,
+                                                       taker_buy_quote_volume]))
+                        safe_ensure_future(self.fill_historical_candles())
+                    elif timestamp_ms > int(self._candles[-1][0]):
+                        # TODO: validate also that the diff of timestamp == interval (issue with 1w, 30d interval).
+                        self._candles.append(np.array([timestamp_ms, open, high, low, close, volume,
+                                                       quote_asset_volume, n_trades, taker_buy_base_volume,
+                                                       taker_buy_quote_volume]))
+                    elif timestamp_ms == int(self._candles[-1][0]):
+                        self._candles.pop()
+                        self._candles.append(np.array([timestamp_ms, open, high, low, close, volume,
+                                                       quote_asset_volume, n_trades, taker_buy_base_volume,
+                                                       taker_buy_quote_volume]))
