+++ hummingbot/data_feed/candles_feed/gate_io_spot_candles/__init__.py
@@ -1,0 +1,3 @@
+from hummingbot.data_feed.candles_feed.gate_io_spot_candles.gate_io_spot_candles import GateioSpotCandles
+
+__all__ = ["GateioSpotCandles"]
