+++ hummingbot/data_feed/candles_feed/gate_io_spot_candles/constants.py
