--- hummingbot/data_feed/candles_feed/gateio_perpetual_candles/__init__.py
@@ -1,3 +1,0 @@
-from hummingbot.data_feed.candles_feed.gateio_perpetual_candles.gateio_perpetual_candles import GateioPerpetualCandles
-
-__all__ = ["GateioPerpetualCandles"]
