--- hummingbot/data_feed/candles_feed/gateio_spot_candles/__init__.py
@@ -1,3 +1,0 @@
-from hummingbot.data_feed.candles_feed.gateio_spot_candles.gateio_spot_candles import GateioSpotCandles
-
-__all__ = ["GateioSpotCandles"]
